Complete initial server generation implementation.
This commit is an initial functional implementation of server generation
that includes some refactoring and restructuring.  The most important
part is the implementation of the router code, which implements the
service layer, including routing requests to the proper API methods and
serialization/deserialization.

Application is renamed to Service, as it includes both the service layer
(largely now implemented by Router) and the application layer.  A service requires a
Router which passes incoming requests to the user implemented API implementation.

For example, this might be the code to instantiate the runner for a service:

```scala
object MyRunner extends Runner {
  def initializeService(environment: Map[String, String], args: Array[String]): Service = {
    val subscriptionApiImpl = new SubscriptionApiImpl()
    SubscriptionApiRouter.createService(subscriptionApiImpl)
  }
}
```

Note that the implementation of the Api class must mix in both the generated Api (shared by
client and server) and an ApiImpl class which provides a shutdown() method, so that the ApiImpl
can handle shutdown of any resources that were initialized in the initializeService method and
passed to the ApiImpl.

This is still preliminary: in particular, handling of path parameters still needs some work to
handle all types and post API endpoints aren't getting passed the path parameters.