Remove shutdown from everything below Server.
This commit removes the shutdown() functions from Service, Router, and the
api implementation classes.  The ApiImpl trait is removed enitrely.

Instead, the initializeService method now returns a tuple of the initialized
service and a shutdown function that is responsible for shutting down any
resources initialized as part of service initialization.

Service is now a trait, since it doesn't need to be a class, although I
suspect we may want to return it to being a class in the future.