Add a simple readiness handler for k8s.
This commit adds a basic readiness handler to the service.

By default, as soon as the user provided initialization function
is completed, the readiness handler will return OK.  (Kubernetes
interprets return code 200-400 as OK)

For more complicated readiness criteria, the user can override a method
in the runner and then set a var `KubernetesHandler.ready` to true
when appropriate.